[Feat] issue-#5 In-game 구현 중 및  dto 일부 수정

+++ src/main/java/likelion/hufsglobal/lgtu/runwithmate/domain/game/dto/GameFinishResDto.java
@@ -16,8 +16,8 @@ public class GameFinishResDto {
     private String winner;
     private String winnerName;
     private String loserName;
-    private Integer pointP1;
-    private Integer pointP2;
-    private Integer dopamineP1;
-    private Integer dopamineP2;
+    private Long pointP1;
+    private Long pointP2;
+    private Long dopamineP1;
+    private Long dopamineP2;
 }
